add phrase detection to SlidingCoherenceDisambiguator; QueryConsole is still missing a way to instantiate a PhraseExtractor.
git-svn-id: file:///home/git-svn-mirror/pushmi-svn-mirror/trunk@65499 dd0e9695-b195-4be7-bd10-2dea1a65a6b6

+++ WikiWord/WikiWord/src/main/java/de/brightbyte/wikiword/model/PhraseOccurance.java
@@ -1,7 +1,6 @@
 package de.brightbyte.wikiword.model;
 
 import java.io.Serializable;
-import java.util.List;
 
 public class PhraseOccurance implements Serializable, Comparable<PhraseOccurance>, TermReference {
 
@@ -45,7 +44,7 @@ public class PhraseOccurance implements Serializable, Comparable<PhraseOccurance
 			return getPhrase();
 		}
 
-		public int getWeight() {
+		public double getWeight() {
 			return weight;
 		}
